move modifyPopupState to correct folder
