+++ packages/desktop/lib/auxiliary/popup/actions/modifyPopupState.ts
