--- packages/desktop/lib/auxiliary/popup/helpers/index.ts
@@ -1,1 +1,0 @@
-export * from './modifyPopupState'
