+++ packages/desktop/views/dashboard/send-flow/views/TransactionSummaryView.svelte
@@ -15,8 +15,7 @@
     } from '@core/wallet/actions'
     import { sendFlowParameters } from '@core/wallet/stores'
     import { getNetworkIdFromSendFlowParameters } from '@core/wallet/utils'
-    import { closePopup } from '@desktop/auxiliary/popup'
-    import { modifyPopupState } from '@desktop/auxiliary/popup/helpers'
+    import { closePopup, modifyPopupState } from '@desktop/auxiliary/popup'
     import { onMount } from 'svelte'
     import { sendFlowRouter } from '../send-flow.router'
     import { EvmTransactionSummary, StardustToEvmTransactionSummary, StardustTransactionSummary } from './components'
